requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
ai21==1.3.0
python-dotenv==1.0.0
pymongo==4.5.0
chromadb==0.4.15
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2